add device_detail
I've only included the attributes that are non-null on a couple of my test devices.

The main motivation for this is to be able to filter devices based
on their location (area/building/floor). There are basically two documented
APIs for that: v1/device-detail and v1/membership. The reason for preferring
the former here in this commit is that the latter is exceptionally slow.

The v1/network-device, contrary to its documentation, does not send location
or locationName data. It always sends nulls for these.